// frontend/server/generated.ts
export type readUsersMeApiV1AuthMeGetResponse200 = {
    data: UserResponse;
    status: 200;
};

export type readUsersMeApiV1AuthMeGetResponseSuccess =
    readUsersMeApiV1AuthMeGetResponse200 & {
        headers: Headers;
    };
export type readUsersMeApiV1AuthMeGetResponse =
    readUsersMeApiV1AuthMeGetResponseSuccess;
// ...
export type getAllChatsApiV1ChatGetResponse200 = {
    data: ChatResponse[];
    status: 200;
};

export type getAllChatsApiV1ChatGetResponseSuccess =
    getAllChatsApiV1ChatGetResponse200 & {
        headers: Headers;
    };
export type getAllChatsApiV1ChatGetResponse =
    getAllChatsApiV1ChatGetResponseSuccess;
// ...
export type getChatApiV1ChatChatIdGetResponse200 = {
    data: ExpandedChatResponse;
    status: 200;
};

export type getChatApiV1ChatChatIdGetResponse422 = {
    data: HTTPValidationError;
    status: 422;
};

export type getChatApiV1ChatChatIdGetResponseSuccess =
    getChatApiV1ChatChatIdGetResponse200 & {
        headers: Headers;
    };
export type getChatApiV1ChatChatIdGetResponseError =
    getChatApiV1ChatChatIdGetResponse422 & {
        headers: Headers;
    };

export type getChatApiV1ChatChatIdGetResponse =
    | getChatApiV1ChatChatIdGetResponseSuccess
    | getChatApiV1ChatChatIdGetResponseError;
// ...
export type getDocumentsForUserApiV1DocumentGetResponse200 = {
    data: DocumentResponse[];
    status: 200;
};

export type getDocumentsForUserApiV1DocumentGetResponseSuccess =
    getDocumentsForUserApiV1DocumentGetResponse200 & {
        headers: Headers;
    };
export type getDocumentsForUserApiV1DocumentGetResponse =
    getDocumentsForUserApiV1DocumentGetResponseSuccess;